Update: handle uncommon linebreaks consistently in rules (fixes #7949)
Currently, rules that deal with line-splitting need to include a list of JS linebreaks in some form (e.g. in a regex character class). As a result, splitting a string into JS lines is implemented in 22 different places in the codebase. As described in https://github.com/eslint/eslint/issues/7949, not all of these implementations are correct (many of them forget to account for \r, \u2028, and \u2029.

This commit updates ast-utils.js with the following properties:

* `astUtils.LINEBREAKS`: a Set containing all JS linebreaks
* `astUtils.LINEBREAK_MATCHER`: a regular expression to match JS linebreaks
* `astUtils.createGlobalLinebreakMatcher`: a function that returns a new global regex to match JS linebreaks

Additionally, it updates the other line-splitting code to use these constants.

Observable changes:

* The error message for `newline-per-chained-call` will handle property names with \u2028 and \u2029 the same way it handles names with \n.
* `no-multi-str` will now report an error for strings that contain \r, \u2028, or \u2029. (Previously, it only reported strings containing \n.)
* `no-unused-vars` will now report the correct location if a `/* global foo */` comment contains a linebreak other than \n. (Previously, it would report an invalid location.)
* `spaced-comment` will no longer report an error if a comment contains tokens in the `exceptions` option followed by \u2028 or \u2029.